renamed simulate.base to simulate.common
